# src-be/tsoa-build/routes.ts
/* tslint:disable */
/* eslint-disable */
// WARNING: This file was auto-generated with tsoa. Please do not modify it. Re-run tsoa to re-generate this file: https://github.com/lukeautry/tsoa
import { Controller, ValidationService, FieldErrors, ValidateError, TsoaRoute, HttpStatusCodeLiteral, TsoaResponse } from '@tsoa/runtime';
// WARNING: This file was auto-generated with tsoa. Please do not modify it. Re-run tsoa to re-generate this file: https://github.com/lukeautry/tsoa
import { RadarDataItemsController } from './../app/controller/radar-data-items.controller';
// WARNING: This file was auto-generated with tsoa. Please do not modify it. Re-run tsoa to re-generate this file: https://github.com/lukeautry/tsoa
import { RadarsController } from './../app/controller/radar.controller';
// WARNING: This file was auto-generated with tsoa. Please do not modify it. Re-run tsoa to re-generate this file: https://github.com/lukeautry/tsoa
import { UserController } from './../app/controller/user.controller';
import { expressAuthentication } from './../app/security';
import * as express from 'express';

// WARNING: This file was auto-generated with tsoa. Please do not modify it. Re-run tsoa to re-generate this file: https://github.com/lukeautry/tsoa

const models: TsoaRoute.Models = {
	RadarDataItemEntity: {
		dataType: 'refObject',
		properties: {
			id: { dataType: 'string', required: true },
			radarId: { dataType: 'string', required: true },
			versionId: { dataType: 'string', required: true },
			name: { dataType: 'string', required: true },
			sector: { dataType: 'string', required: true },
			ring: { dataType: 'string', required: true },
			content: { dataType: 'string', required: true },
		},
		additionalProperties: false,
	},
	// WARNING: This file was auto-generated with tsoa. Please do not modify it. Re-run tsoa to re-generate this file: https://github.com/lukeautry/tsoa
	RadarConfig: {
		dataType: 'refObject',
		properties: {
			name: { dataType: 'string', required: true },
			rings: { dataType: 'array', array: { dataType: 'string' }, required: true },
			sectors: { dataType: 'array', array: { dataType: 'string' }, required: true },
			csv: { dataType: 'string', required: true },
		},
		additionalProperties: false,
	},
	// WARNING: This file was auto-generated with tsoa. Please do not modify it. Re-run tsoa to re-generate this file: https://github.com/lukeautry/tsoa
	RadarEntity: {
		dataType: 'refObject',
		properties: {
			uid: { dataType: 'string', required: true },
			radarId: { dataType: 'string', required: true },
			name: { dataType: 'string', required: true },
			lastUpdatedDate: { dataType: 'string', required: true },
			versionId: { dataType: 'string', required: true },
			rings: { dataType: 'array', array: { dataType: 'string' }, required: true },
			sectors: { dataType: 'array', array: { dataType: 'string' }, required: true },
			config: { ref: 'RadarConfig', required: true },
		},
		additionalProperties: false,
	},
	// WARNING: This file was auto-generated with tsoa. Please do not modify it. Re-run tsoa to re-generate this file: https://github.com/lukeautry/tsoa
	UserRole: {
		dataType: 'refEnum',
		enums: ['admin', 'default'],
	},
	// WARNING: This file was auto-generated with tsoa. Please do not modify it. Re-run tsoa to re-generate this file: https://github.com/lukeautry/tsoa
	UserEntity: {
		dataType: 'refObject',
		properties: {
			fullName: { dataType: 'string', required: true },
			email: { dataType: 'string', required: true },
			role: { ref: 'UserRole', required: true },
			photo: { dataType: 'string', required: true },
		},
		additionalProperties: false,
	},
	// WARNING: This file was auto-generated with tsoa. Please do not modify it. Re-run tsoa to re-generate this file: https://github.com/lukeautry/tsoa
};
const validationService = new ValidationService(models);

// WARNING: This file was auto-generated with tsoa. Please do not modify it. Re-run tsoa to re-generate this file: https://github.com/lukeautry/tsoa

export function RegisterRoutes(app: express.Router) {
	// ###########################################################################################################
	//  NOTE: If you do not see routes for all of your controllers in this file, then you might not have informed tsoa of where to look
	//      Please look into the "controllerPathGlobs" config option described in the readme: https://github.com/lukeautry/tsoa
	// ###########################################################################################################
	app.get('/api/radar-data-items/:radarId', function (request: any, response: any, next: any) {
		const args = {
			radarId: { in: 'path', name: 'radarId', required: true, dataType: 'string' },
		};

		// WARNING: This file was auto-generated with tsoa. Please do not modify it. Re-run tsoa to re-generate this file: https://github.com/lukeautry/tsoa

		let validatedArgs: any[] = [];
		try {
			validatedArgs = getValidatedArgs(args, request, response);
		} catch (err) {
			return next(err);
		}

		const controller = new RadarDataItemsController();

		const promise = controller.getRadar.apply(controller, validatedArgs as any);
		promiseHandler(controller, promise, response, next);
	});
	// WARNING: This file was auto-generated with tsoa. Please do not modify it. Re-run tsoa to re-generate this file: https://github.com/lukeautry/tsoa
	app.get('/api/radars', function (request: any, response: any, next: any) {
		const args = {};

		// WARNING: This file was auto-generated with tsoa. Please do not modify it. Re-run tsoa to re-generate this file: https://github.com/lukeautry/tsoa

		let validatedArgs: any[] = [];
		try {
			validatedArgs = getValidatedArgs(args, request, response);
		} catch (err) {
			return next(err);
		}

		const controller = new RadarsController();

		const promise = controller.getAllLatestRadars.apply(controller, validatedArgs as any);
		promiseHandler(controller, promise, response, next);
	});
	// WARNING: This file was auto-generated with tsoa. Please do not modify it. Re-run tsoa to re-generate this file: https://github.com/lukeautry/tsoa
	app.get('/api/radars/:radarId', function (request: any, response: any, next: any) {
		const args = {
			radarId: { in: 'path', name: 'radarId', required: true, dataType: 'string' },
		};

		// WARNING: This file was auto-generated with tsoa. Please do not modify it. Re-run tsoa to re-generate this file: https://github.com/lukeautry/tsoa

		let validatedArgs: any[] = [];
		try {
			validatedArgs = getValidatedArgs(args, request, response);
		} catch (err) {
			return next(err);
		}

		const controller = new RadarsController();

		const promise = controller.getRadar.apply(controller, validatedArgs as any);
		promiseHandler(controller, promise, response, next);
	});
	// WARNING: This file was auto-generated with tsoa. Please do not modify it. Re-run tsoa to re-generate this file: https://github.com/lukeautry/tsoa
	app.post('/api/radars/:radarId', authenticateMiddleware([{ 'admin-access': [] }]), function (request: any, response: any, next: any) {
		const args = {
			radarId: { in: 'path', name: 'radarId', required: true, dataType: 'string' },
			config: { in: 'body', name: 'config', required: true, ref: 'RadarConfig' },
		};

		// WARNING: This file was auto-generated with tsoa. Please do not modify it. Re-run tsoa to re-generate this file: https://github.com/lukeautry/tsoa

		let validatedArgs: any[] = [];
		try {
			validatedArgs = getValidatedArgs(args, request, response);
		} catch (err) {
			return next(err);
		}

		const controller = new RadarsController();

		const promise = controller.updateRadar.apply(controller, validatedArgs as any);
		promiseHandler(controller, promise, response, next);
	});
	// WARNING: This file was auto-generated with tsoa. Please do not modify it. Re-run tsoa to re-generate this file: https://github.com/lukeautry/tsoa
	app.delete('/api/radars/:radarId', authenticateMiddleware([{ 'admin-access': [] }]), function (request: any, response: any, next: any) {
		const args = {
			radarId: { in: 'path', name: 'radarId', required: true, dataType: 'string' },
		};

		// WARNING: This file was auto-generated with tsoa. Please do not modify it. Re-run tsoa to re-generate this file: https://github.com/lukeautry/tsoa

		let validatedArgs: any[] = [];
		try {
			validatedArgs = getValidatedArgs(args, request, response);
		} catch (err) {
			return next(err);
		}

		const controller = new RadarsController();

		const promise = controller.deleteRadar.apply(controller, validatedArgs as any);
		promiseHandler(controller, promise, response, next);
	});
	// WARNING: This file was auto-generated with tsoa. Please do not modify it. Re-run tsoa to re-generate this file: https://github.com/lukeautry/tsoa
	app.get('/api/user', function (request: any, response: any, next: any) {
		const args = {
			request: { in: 'request', name: 'request', required: true, dataType: 'object' },
			errorResponse: {
				in: 'res',
				name: '401',
				required: true,
				dataType: 'nestedObjectLiteral',
				nestedProperties: { reason: { dataType: 'string', required: true } },
			},
		};

		// WARNING: This file was auto-generated with tsoa. Please do not modify it. Re-run tsoa to re-generate this file: https://github.com/lukeautry/tsoa

		let validatedArgs: any[] = [];
		try {
			validatedArgs = getValidatedArgs(args, request, response);
		} catch (err) {
			return next(err);
		}

		const controller = new UserController();

		const promise = controller.getRadar.apply(controller, validatedArgs as any);
		promiseHandler(controller, promise, response, next);
	});
	// WARNING: This file was auto-generated with tsoa. Please do not modify it. Re-run tsoa to re-generate this file: https://github.com/lukeautry/tsoa

	// WARNING: This file was auto-generated with tsoa. Please do not modify it. Re-run tsoa to re-generate this file: https://github.com/lukeautry/tsoa

	function authenticateMiddleware(security: TsoaRoute.Security[] = []) {
		return (request: any, _response: any, next: any) => {
			let responded = 0;
			let success = false;

			const succeed = function (user: any) {
				if (!success) {
					success = true;
					responded++;
					request['user'] = user;
					next();
				}
			};

			// WARNING: This file was auto-generated with tsoa. Please do not modify it. Re-run tsoa to re-generate this file: https://github.com/lukeautry/tsoa

			const fail = function (error: any) {
				responded++;
				if (responded == security.length && !success) {
					error.status = error.status || 401;
					next(error);
				}
			};

			// WARNING: This file was auto-generated with tsoa. Please do not modify it. Re-run tsoa to re-generate this file: https://github.com/lukeautry/tsoa

			for (const secMethod of security) {
				if (Object.keys(secMethod).length > 1) {
					let promises: Promise<any>[] = [];

					for (const name in secMethod) {
						promises.push(expressAuthentication(request, name, secMethod[name]));
					}

					Promise.all(promises)
						.then((users) => {
							succeed(users[0]);
						})
						.catch(fail);
				} else {
					for (const name in secMethod) {
						expressAuthentication(request, name, secMethod[name]).then(succeed).catch(fail);
					}
				}
			}
		};
	}

	// WARNING: This file was auto-generated with tsoa. Please do not modify it. Re-run tsoa to re-generate this file: https://github.com/lukeautry/tsoa

	function isController(object: any): object is Controller {
		return 'getHeaders' in object && 'getStatus' in object && 'setStatus' in object;
	}

	function promiseHandler(controllerObj: any, promise: any, response: any, next: any) {
		return Promise.resolve(promise)
			.then((data: any) => {
				let statusCode;
				let headers;
				if (isController(controllerObj)) {
					headers = controllerObj.getHeaders();
					statusCode = controllerObj.getStatus();
				}

				// WARNING: This file was auto-generated with tsoa. Please do not modify it. Re-run tsoa to re-generate this file: https://github.com/lukeautry/tsoa

				returnHandler(response, statusCode, data, headers);
			})
			.catch((error: any) => next(error));
	}

	// WARNING: This file was auto-generated with tsoa. Please do not modify it. Re-run tsoa to re-generate this file: https://github.com/lukeautry/tsoa

	function returnHandler(response: any, statusCode?: number, data?: any, headers: any = {}) {
		Object.keys(headers).forEach((name: string) => {
			response.set(name, headers[name]);
		});
		if (data && typeof data.pipe === 'function' && data.readable && typeof data._read === 'function') {
			data.pipe(response);
		} else if (data || data === false) {
			// === false allows boolean result
			response.status(statusCode || 200).json(data);
		} else {
			response.status(statusCode || 204).end();
		}
	}

	// WARNING: This file was auto-generated with tsoa. Please do not modify it. Re-run tsoa to re-generate this file: https://github.com/lukeautry/tsoa

	function responder(response: any): TsoaResponse<HttpStatusCodeLiteral, unknown> {
		return function (status, data, headers) {
			returnHandler(response, status, data, headers);
		};
	}

	// WARNING: This file was auto-generated with tsoa. Please do not modify it. Re-run tsoa to re-generate this file: https://github.com/lukeautry/tsoa

	function getValidatedArgs(args: any, request: any, response: any): any[] {
		const fieldErrors: FieldErrors = {};
		const values = Object.keys(args).map((key) => {
			const name = args[key].name;
			switch (args[key].in) {
				case 'request':
					return request;
				case 'query':
					return validationService.ValidateParam(args[key], request.query[name], name, fieldErrors, undefined, {
						noImplicitAdditionalProperties: 'throw-on-extras',
					});
				case 'path':
					return validationService.ValidateParam(args[key], request.params[name], name, fieldErrors, undefined, {
						noImplicitAdditionalProperties: 'throw-on-extras',
					});
				case 'header':
					return validationService.ValidateParam(args[key], request.header(name), name, fieldErrors, undefined, {
						noImplicitAdditionalProperties: 'throw-on-extras',
					});
				case 'body':
					return validationService.ValidateParam(args[key], request.body, name, fieldErrors, undefined, {
						noImplicitAdditionalProperties: 'throw-on-extras',
					});
				case 'body-prop':
					return validationService.ValidateParam(args[key], request.body[name], name, fieldErrors, 'body.', {
						noImplicitAdditionalProperties: 'throw-on-extras',
					});
				case 'res':
					return responder(response);
			}
		});

		if (Object.keys(fieldErrors).length > 0) {
			throw new ValidateError(fieldErrors, '');
		}
		return values;
	}

	// WARNING: This file was auto-generated with tsoa. Please do not modify it. Re-run tsoa to re-generate this file: https://github.com/lukeautry/tsoa
}

// WARNING: This file was auto-generated with tsoa. Please do not modify it. Re-run tsoa to re-generate this file: https://github.com/lukeautry/tsoa