# src-be/tsoa-build/swagger.json
{
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"RadarDataItemEntity": {
				"properties": {
					"id": {
						"type": "string"
					},
					"radarId": {
						"type": "string"
					},
					"versionId": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"sector": {
						"type": "string"
					},
					"ring": {
						"type": "string"
					},
					"content": {
						"type": "string"
					}
				},
				"required": ["id", "radarId", "versionId", "name", "sector", "ring", "content"],
				"type": "object",
				"additionalProperties": false
			},
			"RadarConfig": {
				"properties": {
					"name": {
						"type": "string"
					},
					"rings": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"sectors": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"csv": {
						"type": "string"
					}
				},
				"required": ["name", "rings", "sectors", "csv"],
				"type": "object",
				"additionalProperties": false
			},
			"RadarEntity": {
				"properties": {
					"uid": {
						"type": "string"
					},
					"radarId": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"lastUpdatedDate": {
						"type": "string"
					},
					"versionId": {
						"type": "string"
					},
					"rings": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"sectors": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"config": {
						"$ref": "#/components/schemas/RadarConfig"
					}
				},
				"required": ["uid", "radarId", "name", "lastUpdatedDate", "versionId", "rings", "sectors", "config"],
				"type": "object",
				"additionalProperties": false
			},
			"UserRole": {
				"enum": ["admin", "default"],
				"type": "string"
			},
			"UserEntity": {
				"properties": {
					"fullName": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"role": {
						"$ref": "#/components/schemas/UserRole"
					},
					"photo": {
						"type": "string"
					}
				},
				"required": ["fullName", "email", "role", "photo"],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "radars-backend",
		"version": "0.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"openapi": "3.0.0",
	"paths": {
		"/api/radar-data-items/{radarId}": {
			"get": {
				"operationId": "GetRadar",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/RadarDataItemEntity"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "radarId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/radars": {
			"get": {
				"operationId": "GetAllLatestRadars",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/RadarEntity"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": []
			}
		},
		"/api/radars/{radarId}": {
			"get": {
				"operationId": "GetRadar",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/RadarEntity"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "radarId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"post": {
				"operationId": "UpdateRadar",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/RadarEntity"
								}
							}
						}
					}
				},
				"security": [
					{
						"admin-access": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "radarId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/RadarConfig"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteRadar",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"security": [
					{
						"admin-access": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "radarId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/user": {
			"get": {
				"operationId": "GetRadar",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UserEntity"
								}
							}
						}
					},
					"401": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"reason": {
											"type": "string"
										}
									},
									"required": ["reason"],
									"type": "object"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": []
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}